minor fixes

+++ core/classes/class.util.php
@@ -476,6 +476,9 @@ class Util
 
         $separator = '========================================================================================' . PHP_EOL;
         foreach ($logs as $log) {
+            if (!file_exists($log)) {
+                continue; // Skip to the next iteration if the log file does not exist
+            }
             $logContent = @file_get_contents($log);
             if ($logContent !== false && !self::endWith($logContent, $separator)) {
                 file_put_contents($log, $separator, FILE_APPEND);
